- Fixed bug in last commit.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2009/imposters@32840 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/engine/3d/impman.cpp
@@ -89,7 +89,10 @@ bool csImposterManager::HandleEvent(iEvent &ev)
       }
     }
 
-    updateQueue.DeleteIndex(i--);
+    if(updateQueue[i]->init || updateQueue[i]->remove)
+    {
+      updateQueue.DeleteIndex(i--);
+    }
   }
 
   return false;
